change spatial storage numeric to float
Using float integrates better with the existing codebase and simplifies
the code. Because the codebase uses floats, we wouldn't gain a lot from
using doubles for the storage, whereas using floats may increase our
storage efficiency. I don't know whether the difference is negligible or
not.

+++ src/util/spatialstore.h
@@ -34,7 +34,7 @@ template <typename T, typename U = u32>
 class SpatialStore
 {
 private:
-	spatial::RTree<double, sp_util::TaggedBBox<U>, 3, 4, 3,
+	spatial::RTree<float, sp_util::TaggedBBox<U>, 3, 4, 3,
 			sp_util::TaggedBBoxIndexable<U>>
 			m_tree;
 	std::map<U, T> m_spacesMap;
